move change_dir to common in tests
